requirements.txt
requests
python-dotenv
GitPython